requirements.txt
# Core Data Science Libraries
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
scipy==1.10.1

# Visualization Libraries
plotly==5.14.1
matplotlib==3.7.2
seaborn==0.12.2

# Web Application
streamlit==1.22.0
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
pymongo==4.6.0
dnspython==2.5.0

# AI/ML
transformers==4.28.0
torch==2.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2

# Feature Engineering
category_encoders==2.6.1
imbalanced-learn==0.10.1

# Statistics and Analysis
statsmodels==0.14.0

# File Handling
openpyxl==3.1.2
xlrd==2.0.1
